[build] adding support for Chinese client up to patch 3 content
Previously, Chinese users using the HD patch would lose spell names, ranks, and descriptions in tooltips, making the game difficult to play with it. Many were playing in the English Client without understanding english in order to use a translator addon to play the game. Adding Chinese was very simple, thus i see no reason why we can't support it, and it does not affect english-speaking users since the change is localized. plus, it does not require maintaining a separate patch. the only side effect MIGHT be slower loading, but we can diagnose this later if necessary.